requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
SQLAlchemy==2.0.40
Werkzeug==2.3.8
Jinja2==3.1.6
qrcode==7.4.2
Pillow==10.0.0
python-dotenv==1.0.1
